// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/filtroAvanceMesaBean.ts
export class FiltroAvanceMesaBean{
  idProceso: number;
  idEleccion: string;
  codigoEleccion: string;
  idAmbito: number;
  ambitoElectoral: string;
  codigoAmbitoElectoral: string;
  idCentroComputo: number;
  codigoCentroComputo: string;
  centroComputo: string;
  departamento: string;
  provincia: string;
  idUbigeo: number;
  mesa: string;
  proceso?: string;
  eleccion?: string;
  schema?: string;
  distrito?:string;
  preferencial?: number;
  acronimo?: string;
}