// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/transversal/enums/tipo-reporte.enum.ts
export enum TipoReporteEnum {
    MesasPorMesa = 1,
    MesasPorEstadoDeActa = 2,
    MesasPorEstadoDeDigitacion = 3
}

export enum CodigoEleccionesEnum {
    COD_ELEC_REV_DIST = "07",
    COD_ELEC_REF_NACIONAL = "05",
    COD_ELEC_DIPUTADOS = "13",
    COD_ELEC_DIST_MULTIPLE = "14",
    COD_ELEC_DIST_UNICO = "15",
    MesasPorEstadoDeDigitacion = 3
}